agentic_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
pymongo==4.10.1
pytest==8.3.3
mongomock==4.2.0.post1
httpx==0.27.2
sentence-transformers==3.0.1
